+++ examples/how-do-i-run-my-own-function-in-the-window-loop.hs
@@ -10,13 +10,14 @@
 #-}
 
 import System.Random
+import Data.Text
 import Control.Concurrent.MVar
 import Language.Javascript.JMacro
 import qualified Graphics.UI.Webviewhs as WHS
 
 main :: IO ()
 main = do
-  count <- newMVar (0 :: Int)
+  counter <- newMVar (0 :: Int)
   WHS.withWindowLoop
     WHS.WindowParams
       { WHS.windowParamsTitle      = "webviewhs - How do I run my own function in the window loop?"
@@ -28,15 +29,19 @@ main = do
       , WHS.windowParamsDebuggable = True
       }
     -- This is the callback JavaScript can execute.
-    (\ _window text -> print text) $
+    (\ _window text -> print text)
+    -- This function runs before the loop.
+    (WHS.WithWindowLoopSetUp    (\ _window -> print ("Setting up." :: Data.Text.Text)))
+    -- This function runs after the loop.
+    (WHS.WithWindowLoopTearDown (\ _window -> print ("Tearing down." :: Data.Text.Text)))
     -- This function runs every window loop.
     -- Return True to continue the loop or False to exit the loop.
-    \ window -> do
-      count' <- takeMVar count
+    $ \ window -> do
+      counter' <- takeMVar counter
       -- Every so often, change the web page background color to a random color.
-      if count' >= 100000
+      if counter' >= 100000
         then do
-          putMVar count 0
+          putMVar counter 0
           red   <- randomRIO (0 :: Int, 255)
           green <- randomRIO (0 :: Int, 255)
           blue  <- randomRIO (0 :: Int, 255)
@@ -54,5 +59,5 @@ main = do
               );
             |]
         else do
-          putMVar count $ count' + 1
+          putMVar counter $ counter' + 1
           return True