+++ examples/how-do-i-run-some-function-in-the-main-window-thread-while-in-some-other-thread.hs
@@ -11,6 +11,7 @@
 
 import System.Random
 import Data.Maybe
+import Data.Text
 import Control.Monad
 import Control.Concurrent
 import Language.Javascript.JMacro
@@ -18,7 +19,7 @@ import qualified Graphics.UI.Webviewhs as WHS
 
 main :: IO ()
 main = do
-  count    <- newMVar (0 :: Int)
+  counter  <- newMVar (0 :: Int)
   continue <- newMVar True
   WHS.withWindowLoop
     WHS.WindowParams
@@ -31,15 +32,19 @@ main = do
       , WHS.windowParamsDebuggable = True
       }
     -- This is the callback JavaScript can execute.
-    (\ _window text -> print text) $
+    (\ _window text -> print text)
+    -- This function runs before the loop.
+    (WHS.WithWindowLoopSetUp    (\ _window -> print ("Setting up." :: Data.Text.Text)))
+    -- This function runs after the loop.
+    (WHS.WithWindowLoopTearDown (\ _window -> print ("Tearing down." :: Data.Text.Text)))
     -- This function runs every window loop.
     -- Return True to continue the loop or False to exit the loop.
-    \ window -> do
-      count' <- takeMVar count
+    $ \ window -> do
+      counter' <- takeMVar counter
       -- Every so often, change the web page background color to a random color.
-      if count' >= 100000
+      if counter' >= 100000
         then do
-          putMVar count 0
+          putMVar counter 0
           -- Instead of changing the background color in the main thread,
           -- we'll change it from another thread by dispatching the
           -- background-changing function to the window's main thread.
@@ -69,5 +74,5 @@ main = do
                 void $ tryPutMVar continue success
           fromMaybe True <$> tryTakeMVar continue
         else do
-          putMVar count $ count' + 1
+          putMVar counter $ counter' + 1
           return True